More splitting out of the directshow code
